Remove the lock (#8)
The `eos_lock` cannot protect the eos against concurrent changes, as there are multiple processes on different servers, which do not share the lock. That is handled by an API call creating a lock for the process.

The lock was required as there `admin_context`  with corresponding database session, which was shared between threads. The underlying db api may not be thread-safe (e.g. psycopg2)
The proper locking should have been in the `networking_arista.common.db_lib.NeutronNets` class, but the removal of a "global" shared admin_context eliminated the need for locking.