Refresh port-channel mappings and save used ids
We're still caching the port -> port-channel mapping, but refresh the
whole list with a single command listing all port-channel, when a port
cannot be found in the channel-mapping.

Listing all configure port-channels lets us create a list of already
used port-channel ids. We need this list later on, when we want to
create port-channels on-the-fly instead of statically configuring them
once.