Fix segmentation_id attr-item duality
Sometimes segmentation_id is an item of a dict, sometimes it is an
attr. Most likely it is an item in plug_into_network() and an attr
in unplug_from_network(), but to be on the safe side we're going to
check for an attr, and, if not present, return the item.