Fix remote security groups with borrowed neutron func
Originally in this fork NeutronNets became a SecurityGroupServerRpcMixin
to get the membership of servers to a remote security group.
With neutron > mitaka this class also implements a notifier interface,
which we currently do not support. There are better ways to get
the remote security group memberships, but for now we fix this by no
longer inheriting from the RpcMixin and copy the required function
over in our code.