Don't trust tooz callbacks, always check if leader
We have observed that often tooz does not prevent two instances of
the driver making changes to the device, which most likely is caused
by _become_leader() only being called by tooz when the current node
is elected as leader, but not when it is not any more (due to a tooz
timeout for example).

Therefore we now no longer trust the _is_leader flag, but do the
quick memcache lookup for the current leader once for each sync
cycle.