Fix leadership election by increasing timeouts
The default leader_timeout is 30 seconds, while the default sync
interval is 60 seconds. This results in the leader timing out after
every single sync. By setting leader_timeout to the same timeout as
membership_timeout this should no longer happen.

lock_timeout is increased out of consistency reason. There is nothing
at the moment in the driver or tooz that is relying on this value
being set.