Fix interface --> portchannel association in sg syncloop
This commit fixes the security group association on port-channels in the
regular syncloop. To figure out the mapping between interfaces and
port-channels _get_interface_membership() gets called. Before this
commit the port_security_groups keys were passed as an iterator, which
is consumed two times, resulting in an empty mapping. This no longer
happens when using keys().

>>> self._get_interface_membership(server, port_security_groups.iterkeys())
{}
>>> self._get_interface_membership(server, port_security_groups.keys())
{u'Ethernet2': u'Port-Channel102',
 u'Ethernet3': u'Port-Channel103',
 u'Ethernet4': u'Port-Channel104',
 u'Ethernet5': u'Port-Channel105',
 u'Ethernet6': u'Port-Channel106',
 u'Ethernet7': u'Port-Channel107',
 u'Ethernet9': u'Port-Channel109'}