Only deprovision vlan if no port on device has this network
This is again a special circumstance of our environment. Each server is
connected via a single mlag to our switches. To prevent vlans being
removed that are still in use, we now check that on removal there is no
port left for the device that this port belonged to that is also in this
network.