On port unplug remove vlan from ALL ports
As we now have protection in place that a vlan is only decommissioned
when the network is removed from all host ports we also need to make
sure we now remove the vlan from all ports.

When the last port is unplugged we now look for its portchannel and from
there find all ports belonging to this pc, both on the switch the port
is mapped to and also on the mlag peer.