Prevent loopingcalls to run in multiple processes
The Arista mechanism driver has multiple loopingcalls taking care of
syncing security groups and saving switch config. When neutron forks off
its workers it seems these looping calls are present in all workers, but
are only running in one and paused in the others. Due to some unknown
cause these sleeping threads might somehow get resumed after an unknown
amount of time (hours, days, never) and then we are in a situation where
every single worker on the active driver host is hammering the API,
essentially bringing down the switch API with this.

For the synchronization we generate a member id which includes the PID.
For the forked processes this PID will be different, resulting in a
different generated id for the coordinator. To prevent these threads
from running we now check if the coordinator id is differing from what
the coordinator was set up with. If it differs we shut down the thread.

Suspected causes for this are some sort of failure inside the Tooz
coordinator or something causing a ToozError (not a
ToozConnectionError). Other exceptions might be able to trigger this as
well, but could not be reproduced yet.