Make code compatible with python3
Whatever 2to3 has to offer! Also we replace a couple of __nonzero__()
calls with __bool__() as this is called by python3, making those test
succeed on python3, but fail on python2. So we're essentially beginning
to drop python2 support here.

Python2 allowed list of dicts to be sorted by whatever metric. We assume
only a stable sort order is desired, so in some places we remove it (as
python3.6 has a stable insertion order), in other places we simply sort
by sorted list of keys.