Fix db deadlock in _cleanup_db() by adding retry annotation
When starting multiple neutron-server instances at ones, the complex query caused
by _cleanup_db is holding a lock for quite a long time, causing other instances to fail on startup
since the query is a mandatory step in the initalization.

This patch first adds a db retry annotation that will retry in case of a detected deadlock.