Fix _cleanup_db by correctly deleting VM entries if neutron port is missing.
This commit changes the db cleanup of arista driver to cleanup arista_vms entries based
on deleted neutron ports, and cleanup arist_nets based on deleted neutron networks.

The old code relied on the network to be deleted for arista_vms as a condition, which caused a huge amount of orphaned
arista_vms - But baremetal instances could be deleted and recreated any time without network deletion - causing a leak.