Remove broken config validation for nocvx
NoCVX does not need eapi_host or eapi_username to be set, but there was
still code present checking for the presence of these two values. As
NoCVX is not using these two variables, we don't require them to be
present and can just override the checking method with a NOOP method.