Don't deny network deletion on still provisioned ports
Sometimes networking-arista puts ports in its DB that it shouldn't have,
as they never where provisioned there. We have not spent much time in
figuring out why that is due to time limitations. All we want this
driver to do currently is configure the baremetal Arista switches that
we have and to NOT block network deletions.

Therefore I am now removing the part that raises an error in the
delete_network_precommit() hook and replace it by a piece of code that
just deletes all remaining ports for the network in the Arista shadow
db. Note that we only touch the Arista shadow tables here, which are
only used by the driver (and even there with nocvx I'm not entirely sure
that they are used by us). Upstream has dropped them a long time ago,
but we didn't spent the time to actually clean up the whole driver.