Prevent MemberAlreadyExist when pod is being restarted
If the pod is being restarted this will often lead to it crashlooping.
That is, as the we generate the coordinator member_id on the PID and the
hostname. Yet, that will be the same when the pod is being restarted. As
long as the interval between restarts is shorter than
`membership_timeout` a restart will be unsuccessful as the old member
has not timed out yet.