Use the new enginefacade from oslo_db
This is a hacky implementation of engine facade usage. I've more or less
added the CONTEXT_READER/CONTEXT_WRITER everywhere where we previously
created a transaction, where called get_reader_session() /
get_writer_session() or where a session was used. This might not be
necessary everywhere, but as we run this driver only in maintenance mode
until end of the year it hopefully will do.