More reader contexts for own db lib
This driver does some checks in its postcommit hooks that query the
database. These hooks run outside of a DB transaction, which, under
certain conditions can lead to an open transaction not being commited.
This happens, because Neutrons own db methods don't open a transaction
themselves. Therefore we now add a CONTEXT_READER decorator to all
get methods of the NeutronNets class.

In this specific case we had problems with update_port_postcommit(), in
which _network_owner_tenant() is called, which then tries to get a
network by its id outside of a transaction. This led to the quota driver
raising the following exception:

Method <function remove_reservation at 0x7fac6bb65c60> cannot be called within a transaction

The other option of solving this would be to figure out all places where
theses methods are being used and take care of opening a transaction
there, but as this driver will soon be removed we won't put in the
effort.